+++ examples/2-D_discrete_x.py
@@ -24,23 +24,27 @@
 from pyharmonysearch import ObjectiveFunctionInterface, harmony_search
 from math import pow
 import random
+from bisect import bisect_left
+from multiprocessing import cpu_count
 
 class ObjectiveFunction(ObjectiveFunctionInterface):
 	"""
-		This is a toy objective function that contains only continuous variables.
+		This is a toy objective function that contains a mixture of continuous and discrete variables.
 
 		Goal:
 
 			maximize -(x^2 + (y+1)^2) + 4
 			The maximum is 4 at (0, -1).
 
-		Note that since all variables are continuous, we don't actually need to implement get_index() and get_num_discrete_values().
+		In this implementation, x is a discrete variable with choices ranging from -100 to 100 in increments of 1.
+		y is still a continuous variable.
 	"""
 
 	def __init__(self):
-		self._lower_bounds = [-1000, -1000]
-		self._upper_bounds = [1000, 1000]
+		self._lower_bounds = [None, -1000]
+		self._upper_bounds = [None, 1000]
 		self._variable = [True, True]
+		self._discrete_values = [[x for x in xrange(-100, 101)], None]
 		
 		#define all input parameters
 		self._maximize = True #do we maximize or minimize?
@@ -49,43 +53,62 @@ class ObjectiveFunction(ObjectiveFunctionInterface):
 		self._hmcr = 0.75 #harmony memory considering rate
 		self._par = 0.5 #pitch adjusting rate
 		self._mpap = 0.25 #maximum pitch adjustment proportion (new parameter defined in pitch_adjustment()) - used for continuous variables only
-		self._mpai = 2 #maximum pitch adjustment index (also defined in pitch_adjustment()) - used for discrete variables only
-		self._random_seed = 8675309 #optional random seed for reproducible results
+		self._mpai = 10 #maximum pitch adjustment index (also defined in pitch_adjustment()) - used for discrete variables only
 
 	def get_fitness(self, vector):
-		"""
-			maximize -(x^2 + (y+1)^2) + 4
-			The maximum is 4 at (0, -1).
-		"""
 		return -(pow(vector[0], 2) + pow(vector[1] + 1, 2)) + 4
 
-	def get_value(self, i, index=None):
-		"""
-			Values are returned uniformly at random in their entire range. Since both parameters are continuous, index can be ignored.
-		"""
+	def get_value(self, i, j=None):
+		if self.is_discrete(i):
+			if j:
+				return self._discrete_values[i][j]
+			return self._discrete_values[i][random.randint(0, len(self._discrete_values[i]) - 1)]
 		return random.uniform(self._lower_bounds[i], self._upper_bounds[i])
 
 	def get_lower_bound(self, i):
+		"""
+			This won't be called except for continuous variables, so we don't need to worry about returning None.
+		"""
 		return self._lower_bounds[i]
 	
 	def get_upper_bound(self, i):
+		"""
+			This won't be called except for continuous variables.
+		"""
 		return self._upper_bounds[i]
+
+	def get_num_discrete_values(self, i):
+		if self.is_discrete(i):
+			return len(self._discrete_values[i])
+		return float('+inf')
+	
+	def get_index(self, i, v):
+		"""
+			Because self.discrete_values is in sorted order, we can use binary search.
+		"""
+		return ObjectiveFunction.binary_search(self._discrete_values[i], v)
+		
+	@staticmethod
+	def binary_search(a, x):
+		"""
+			Code courtesy Python bisect module: http://docs.python.org/2/library/bisect.html#searching-sorted-lists
+		"""
+		i = bisect_left(a, x)
+		if i != len(a) and a[i] == x:
+			return i
+		raise ValueError
 	
 	def is_variable(self, i):
 		return self._variable[i]
 	
 	def is_discrete(self, i):
-		#all variables are continuous
-		return False
+		return self._discrete_values[i] is not None
 	
 	def get_num_parameters(self):
 		return len(self._lower_bounds)
 	
 	def use_random_seed(self):
 		return hasattr(self, '_random_seed') and self._random_seed
-	
-	def get_random_seed(self):
-		return self._random_seed
 	
 	def get_max_imp(self):
 		return self._max_imp
@@ -110,6 +133,6 @@ class ObjectiveFunction(ObjectiveFunctionInterface):
 
 if __name__ == '__main__':
 	obj_fun = ObjectiveFunction()
-	num_processes = 1
-	num_iterations = 1 #because random_seed is defined, there's no point in running this multiple times
+	num_processes = cpu_count() #use number of logical CPUs
+	num_iterations = num_processes * 5 #each process does 5 iterations
 	print harmony_search(obj_fun, num_processes, num_iterations)