+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+## 1.0.1 (2013-06-30)
+* Fixed a bug in the pitch adjustment step.
+* Added a new, more complicated, example that solves a 5-D linear system of equations.
+
 ## 1.0 (2013-06-30)
 * Refactored code so that HS is now implemented as a class called `HarmonySearch`. A non-class method `harmony_search` is used to call `HarmonySearch` with multiple processes. This now means we can do `from pyharmonysearch import ObjectiveFunctionInterface, harmony_search` when implementing an objective function.
 * Added `setup.py` so that pyHS can be installed and imported anywhere.